fastapi/requirements.txt
fastapi
uvicorn
asyncpg
pydantic
python-dotenv
requests
boto3